requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.6.1
argon2-cffi==23.1.0
pyjwt==2.8.0
python-multipart==0.0.9
pydantic-settings
python-jose[cryptography]